% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffinity.R
\name{getAffinity}
\alias{getAffinity}
\title{Predict the binding affinity between drugs and proteins}
\usage{
getAffinity(
  smilesDF,
  seqDF,
  DRN,
  condaenv,
  path,
  pretrained_model = "MPNN_CNN_BindingDB_IC50"
)
}
\arguments{
\item{smilesDF}{A data frame with "drug" column as the drug name and "SMILES" column as the simplified molecular input line entry system format of the drug.}

\item{seqDF}{A data frame with "protein" column as symbols of proteins and "Sequence" column as the amino acid sequences of the proteins.}

\item{DRN}{A data frame contains two columns: "drug"(drug name) and "protein"(symbol).}

\item{condaenv}{The virtual environment path for DeepPurpose installed using conda.}

\item{path}{The path where the DeepPurpose package is located.}

\item{pretrained_model}{The DeepPurpose model type used, which defaults to 'MPNN_CNN_BindingDB_IC50'.}
}
\description{
The function uses the DeepPurpose tool to predict the binding affinity between drug-protein pairs,
    and the results are saved in the "result/virtual_screening.txt" file.
}
\examples{
\dontrun{
  data(drug_smiles)
  data(protein_sequences)
  data(DRN)
  getAffinity(smiles,seqs,DRN,condaenv="C:/Users/A/.conda/envs/DeepPurpose",
    path="C:/Users/A/DeepPurpose")
}

}
