% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDRN.R
\name{getDRN}
\alias{getDRN}
\title{Construct the drug response network}
\usage{
getDRN(moduleDF, drugDF)
}
\arguments{
\item{moduleDF}{A data frame storing expression values of the module, with rows representing proteins and columns representing samples.}

\item{drugDF}{A data frame storing drug sensitivity scores predicted by the module, with rows representing samples and columns representing drugs.}
}
\value{
a list containing the drug response network and its node information
}
\description{
This function utilizes protein expression within modules and predicted drug sensitivity scores based on
   module expression profiles to compare scores between two groups with high and low expression of a particular protein.
   If there is a significant difference (p < 0.05), it suggests a potential interaction between the drug and the protein.
}
\examples{
\dontrun{
  drugDF <- read.csv("calcPhenotype_Output/DrugPredictions.csv",row.names = 1,check.names = F)
  getDRN(moduleDF,drugDF)
}

}
