% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModule.R
\name{getModule}
\alias{getModule}
\title{Identify subtype-specific network modules}
\usage{
getModule(expr_file, group_file, subtype, ppi_file)
}
\arguments{
\item{expr_file}{Expression profile without log2 transformation, with rows representing proteins and columns representing samples.}

\item{group_file}{Subtype information, with the first column being samples and the second column being subtype grouping.}

\item{subtype}{A vector representing the analysis of a specific subtype.}

\item{ppi_file}{Interactome data used to construct networks.}
}
\value{
This function doesn't return anything, but saves the results to the "NetSDR_results/Modules" files.
}
\description{
Calculate subtype-specific signature proteins based on the proteome and map them
    onto the human interactome (combined score >0.4) to identify network modules for the subtype.
    The signature proteins are saved in the 'signature.csv' file. The 'edges.txt' file stores
    the subtype-specific network, while the 'node_Module.txt' and 'edge_Module.txt' files provide
    information on the nodes and edges of robust modules, respectively.
}
\examples{
\dontrun{
  expr_file <- "inst/extdata/expression.txt"
  group_file <- "inst/extdata/group.txt"
  subtype <- "G-IV"
  ppi_file <- "inst/extdata/ppi_String400.txt"
  getModule(expr_file,group_file,subtype,ppi_file)
}

}
