% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleResponse.R
\name{getModuleResponse}
\alias{getModuleResponse}
\title{Calculate the drug response levels of the modules}
\usage{
getModuleResponse(moduleDF)
}
\arguments{
\item{moduleDF}{A data frame storing expression values of the module, with rows representing proteins and columns representing samples.}
}
\value{
The number of significantly affected and unaffected drugs.
}
\description{
This function predicts the clinical drug response of patients with a specific subtype
    based on the expression profile of the module, and calculates the number of drugs whose
    sensitivity score is affected by the expression of the module.
}
\examples{
\dontrun{
  sig_count <- getModuleResponse(moduleDF)
}

}
