% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROC.R
\name{runROC}
\alias{runROC}
\title{Perform ROC analysis}
\usage{
runROC(pred_DPI)
}
\arguments{
\item{pred_DPI}{A data frame contains DPIs and their ps, where the first column represents drugs, the second column represents proteins, and the third column represents the ps value.}
}
\value{
A data frame storing AUC from different cut-off.
}
\description{
This function performs ROC validation on the predicted DPI. Extract DPIs greater than or equal to a certain ps in sequence,
    allocate 1 (known) and 0 (unknown) according to the drug-target interactions of TTD and DrugBank, and perform ROC analysis.
}
\examples{
\dontrun{
  runROC(pred_DPI)
}

}
