% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffinity.R
\name{getAffinity}
\alias{getAffinity}
\title{Predict the binding affinity between drugs and proteins}
\usage{
getAffinity(
  smiles_file,
  seq_file,
  DRN_file,
  py_env,
  pretrained_model = "MPNN_CNN_BindingDB_IC50"
)
}
\arguments{
\item{smiles_file}{The information of simplified molecular input line entry system format of the drug, where the first column is the name or ID of the drug,
and the second column is its smiles symbol.}

\item{seq_file}{The amino acid sequences of proteins, with the first column representing the protein symbol and the second column representing the sequence.}

\item{DRN_file}{Drug-protein interaction data, contains two columns: the first column is "protein" (symbol), and the second column is "drug" (drug name).}

\item{py_env}{The virtual environment path for python installed using conda.}

\item{pretrained_model}{The DeepPurpose model type used, which defaults to 'MPNN_CNN_BindingDB_IC50'.}
}
\description{
The function uses the DeepPurpose tool to predict the binding affinity between drug-protein pairs,
    and the results are saved in the "result/virtual_screening.txt" file.
}
\examples{
\dontrun{
  smiles_file <- "inst/extdata/smiles.txt"
  seq_file <- "inst/extdata/sequences.txt"
  DRN_file <- "inst/extdata/DRN.txt"
  py_env <- "D:/Software/anaconda3/envs/py3.9"
  getAffinity(smiles_file,seq_file,DRN_file,py_env)
}

}
