% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDRN.R
\name{getDRN}
\alias{getDRN}
\title{Construct the drug response network}
\usage{
getDRN(expr_file)
}
\arguments{
\item{expr_file}{Expression profile of the module, with rows representing proteins and columns representing samples.}
}
\value{
This function doesn't return anything, but saves the results to the "NetSDR_results/DRN" files.
}
\description{
This function utilizes protein expression within modules and predicted drug sensitivity scores based on
   module expression profiles to compare scores between two groups with high and low expression of a particular protein.
   If there is a significant difference (p < 0.05), it suggests a potential interaction between the drug and the protein.
}
\examples{
\dontrun{
  expr_file <- "inst/extdata/expression_module.txt"
  getDRN(expr_file)
}

}
