% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModule.R
\name{getModule}
\alias{getModule}
\title{Identify subtype-specific network modules}
\usage{
getModule(exprDF, groupDF, subtype)
}
\arguments{
\item{exprDF}{A data frame storing expression values, with rows representing proteins and columns representing samples.}

\item{groupDF}{A data frame storing subtype information, with the first column being samples and the second column being subtype grouping.}

\item{subtype}{A vector representing the analysis of a specific subtype.}
}
\value{
This function doesn't return anything, but saves the results to the specified files.
}
\description{
Calculate subtype-specific signature proteins based on the proteome and map them
    onto the human interactome (combined score >0.4) to identify network modules for the subtype.
    The signature proteins are saved in the 'signature.csv' file. The 'edges.txt' file stores
    the subtype-specific network, while the 'node_Module.txt' and 'edge_Module.txt' files provide
    information on the nodes and edges of robust modules, respectively.
}
\examples{
\dontrun{
  data(Expr_Group)
  getModule(exprDF=Expr,groupDF=Group,subtype="G-IV")
}

}
