% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleResponse.R
\name{getModuleResponse}
\alias{getModuleResponse}
\title{Calculate the drug response levels of the modules}
\usage{
getModuleResponse(expr_file)
}
\arguments{
\item{expr_file}{Expression profile of the module, with rows representing proteins and columns representing samples.}
}
\value{
This function doesn't return anything, but saves the results to the "NetSDR_results/DRN" files.
}
\description{
This function predicts the clinical drug response of patients with a specific subtype
    based on the expression profile of the module, and calculates the number of drugs whose
    sensitivity score is affected by the expression of the module.
}
\examples{
\dontrun{
  expr_file <- "inst/extdata/expression_module.txt"
  getModuleResponse(expr_file)
}

}
