% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPS.R
\name{getPS}
\alias{getPS}
\title{Calculat drug score(ps) using PRS methods for DRN.}
\usage{
getPS(BA_file, PPIN_file, py_env)
}
\arguments{
\item{BA_file}{The path to the predicted binding affinity file.}

\item{PPIN_file}{The path to the PPI network file (or to a certain module).}

\item{py_env}{The virtual environment path for python installed using conda.}
}
\value{
A data frame contains the ps of DPI.
}
\description{
This function uses the PRS algorithm to calculate the ps of DPIs and saves it in the "prs_dti_score.csv" file.
}
\examples{
\dontrun{
  py_env <- "D:/Software/anaconda3/envs/py3.9"
  BA_file <- "inst/extdata/virtual_screening.txt"
  PPIN_file <- "inst/extdata/edge_module.txt"
  getPS(BA_file,PPIN_file,py_env)
}

}
